moved backend weekconverter to weekExtenstions
